Move RQDLogger into opencue.cuelogging
